Update All Stripe Products Script
Created a script that can be used to update the products on Stripe to match Supabase. This updates exchange rates, creates new products where they don't yet exist on Stripe, and deactivates old products.

This means the checkout will be able to load faster because much of the work is done in advance and cached for later. So when users go to the checkout page, they won't have to wait for all of the products in their cart to update on Stripe (which can introduce a few seconds of latency).

+++ netlify/functions/updateStripeProducts.mts
@@ -2,11 +2,13 @@ import {Context} from "@netlify/functions";
 import type {ProductData, RawProductData, WebhookPayload} from "@shared/types/supabaseTypes.ts";
 import {VALIDATORS} from "@shared/schemas/schemas.ts";
 import {NetlifyFunctionError} from "@shared/errors.ts";
-import {supabaseAnon} from "../lib/getSupabaseClient.mts";
-import {stripe} from "../lib/stripeObject.mts";
+import {supabaseAnon} from "../lib/getSupabaseClient.ts";
+import {stripe} from "../lib/stripeObject.ts";
 import Stripe from "stripe";
 import {getProducts} from "@shared/functions/supabaseRPC.ts";
-import {generateStripeCurrencyOptions, getImageURL} from "../lib/lib.mts";
+import {generateStripeCurrencyOptions, getImageURL} from "../lib/lib.ts";
+import {supabase} from "../../src/lib/supabaseRPC.tsx";
+import {ConversionRates} from "@shared/functions/price.ts";
 
 /**
  * Sync saved Stripe Product data with Supabase
@@ -23,15 +25,16 @@ import {generateStripeCurrencyOptions, getImageURL} from "../lib/lib.mts";
  * just the current default. This isn't done by default to save on API calls and resources since it shouldn't be
  * necessary more than once.
  *
+ * Specify `sku` as a Supabase product SKU to request that a specific SKU be updated. This only works if the body is
+ * empty and the request is local (i.e. running in Netlify Local Development Environment). This is only designed to be
+ * used in development and is not an available option in production.
+ *
  * @endpoint POST /.netlify/functions/updateStripeProducts
  * @see [Supabase Documentation](https://supabase.com/docs/guides/database/webhooks)
  * @see [Webhook Configuration](https://supabase.com/dashboard/project/iumlpfiybqlkwoscrjzt/integrations/webhooks/webhooks)
  * @see WebhookPayload
  */
-export default async function handler(request: Request, _context: Context): Promise<Response> { try {
-    // TODO: Figure out how to get round the 30s time out when updating all products. Possibly just run a script locally
-    //  since I'll only need to update all products myself so can do it without running through Netlify?
-
+export default async function handler(request: Request, context: Context): Promise<Response> { try {
     // Check form of request
     if (request.method !== "POST" && request.method !== "GET") {
         console.error(`Method ${request.method} not allowed`)
@@ -40,60 +43,80 @@ export default async function handler(request: Request, _context: Context): Prom
         console.error("'Content-Type' must be 'application/json'")
         return new Response("'Content-Type' must be 'application/json'", {status: 412})
     }
-    const payload = await getBody(request);
+
+    // Authorise request.
+    // Key is static so this isn't a hard and fast check, should still be careful with request contents.
+    if (request.headers.get("Authorization") != `Bearer ${process.env.SUPABASE_WEBHOOK_SIGNING_SECRET}`) {
+        return new Response(undefined, {status: 403})
+    }
+
+    // Check if request was called locally.
+    // Locally, this function can be run to update products without need for a change to the products table.
+    let local = context.ip === "::1"
+
+    const body = await request.json()
     const searchParams = new URL(request.url).searchParams
-    switch (payload?.type) {
+    const archiveAllOldPrices = searchParams.has("archiveAllOldPrices")
+    let payload // Supabase Webhook Payload (If this is a webhook request)
+    switch (body.type) {
         case "INSERT": // Add a new product on Stripe
+            payload = await processBody(request);
             await handleInsertCase(payload.record)
             break;
         case "UPDATE": // Update product data on Stripe
-            await handleUpdateCase(payload.record)
+            payload = await processBody(request);
+            await handleUpdateCase(payload.record, archiveAllOldPrices)
             break;
         case "DELETE": // Remove a product from Stripe
+            payload = await processBody(request);
             await handleDeleteCase(payload.old_record);
             break;
-        default: // No payload, update all products
-            await updateAllProducts();
+        default: // Payload not from Supabase, if local, ignore this and update the product from query string.
+            if (local && searchParams.has("sku")) {
+                // TODO: Handle deleting old products on Stripe too.
+                const sku = Number(searchParams.get("sku")!);
+                if (!Number.isNaN(Number(sku))) {
+                    // Local request means the body contains exchange rates
+                    await updateFromSku(
+                        sku,
+                        searchParams.has("archiveAllOldPrices"),
+                        body
+                    )
+                }
+            } else if (local) {
+                return new Response(
+                    "Local request and no body provided, but no `sku` query string parameter was provided.",
+                    {status: 400}
+                );
+            } else {
+                return new Response(
+                    "Invalid Payload",
+                    {status: 400}
+                )
+            }
             break;
     }
 
-
     return new Response(undefined)
+
 } catch(error: unknown) {
     if (VALIDATORS.NetlifyFunctionError(error)) {
         const netlifyFunctionError = error as NetlifyFunctionError;
         return new Response(netlifyFunctionError.message, {status: netlifyFunctionError.status});
     } else {
+        console.error(error)
         return new Response(undefined, {status: 500})
     }
 }}
 
 /**
- * Extract the body of the request if it exists
+ * Validate the body of the request and return a typed version of the body.
  */
-async function getBody(request: Request): Promise<WebhookPayload | undefined> {
-    let webhookPayload: WebhookPayload;
-    const bodyText = await request.text() // Have to get text first since Supabase doesn't update bodyUsed param.
-    if (bodyText) {
-        const body: unknown = JSON.parse(bodyText)
-        if (!VALIDATORS.WebhookPayload(body)) {
-            throw new NetlifyFunctionError("Request body malformed", 400);
-        }
-        webhookPayload = body as WebhookPayload;
-        return webhookPayload;
+async function processBody(body: any): Promise<WebhookPayload> {
+    if (!VALIDATORS.WebhookPayload(body)) {
+        throw new NetlifyFunctionError("Request body malformed", 400);
     }
-}
-
-/** Fetch active product data from Stripe.
- * @returns An array of objects containing all active products on Stripe.
- */
-async function fetchActiveStripeProducts() {
-    const prods: Stripe.Product[] = []
-    await stripe
-        .products
-        .list({active: true})
-        .autoPagingEach((prod) => {prods.push(prod)})
-    return prods;
+    return body as WebhookPayload;
 }
 
 /**
@@ -101,11 +124,11 @@ async function fetchActiveStripeProducts() {
  * @param supabaseSKU The SKU of the product to find a match for.
  * @return The Stripe product, or undefined if no product was able to be found.
  */
-async function fetchStripeProduct(supabaseSKU: number): Promise<Stripe.Product | undefined> {
+export async function fetchStripeProduct(supabaseSKU: number): Promise<Stripe.Product | undefined> {
     // Try to find product by URL first
     const prods = await stripe
         .products
-        .list({url: `https://thisshopissogay.com/products/${supabaseSKU}`})
+        .list({url: `https://thisshopissogay.com/products/${supabaseSKU}`, active: true})
     if (prods.data.length > 0) {return prods.data[0]}
 
     // Products which are running on the outdated per-checkout update flow may not have a URL set. Fetch all products
@@ -124,11 +147,13 @@ async function fetchStripeProduct(supabaseSKU: number): Promise<Stripe.Product |
  * @param supabaseProduct The data to update the Stripe product with.
  * @param archiveAllOldPrices Whether to archive all the old prices, instead of just the current default, when updating
  * prices.
+ * @param cachedExchangeRates Cached conversion rates to prevent having to fetch new ones
  */
 async function updateStripeProduct(
     stripeProduct: Stripe.Product,
     supabaseProduct: ProductData,
-    archiveAllOldPrices = false
+    archiveAllOldPrices = false,
+    cachedExchangeRates?: ConversionRates
 ) {
     console.log(`Updating ${supabaseProduct.name} on Stripe...`)
 
@@ -137,7 +162,9 @@ async function updateStripeProduct(
     const lookup_key = `${supabaseProduct.sku}_${supabaseProduct.name}`
     const currency_options = await generateStripeCurrencyOptions(
         Math.round(supabaseProduct.price*100),
-        "PRODUCT"
+        "PRODUCT",
+        undefined,
+        cachedExchangeRates
     )
     process.stdout.write("Updating Stripe price data - Saving new price data... ")
     const newPrice = await stripe.prices.create({ // Create new price
@@ -158,7 +185,11 @@ async function updateStripeProduct(
         description: supabaseProduct.description,
         name: supabaseProduct.name,
         images: imageURL ? [imageURL] : [],
-        metadata: {category_id: supabaseProduct.category.id, category: supabaseProduct.category.name},
+        metadata: {
+            category_id: supabaseProduct.category.id,
+            category: supabaseProduct.category.name,
+            sku: supabaseProduct.sku
+        },
         url: `https://thisshopissogay.com/products/${supabaseProduct.sku}`,
         default_price: newPrice.id // Set default price to the newly created one
     })
@@ -189,15 +220,18 @@ async function updateStripeProduct(
 /**
  * Create a new product on Stripe with the data from the given Supabase product.
  * @param supabaseProduct The data to create the new Stripe product to match.
+ * @param cachedExchangeRates Cached conversion rates to prevent having to fetch new ones.
  */
-async function createStripeProduct(supabaseProduct: ProductData) {
+async function createStripeProduct(supabaseProduct: ProductData, cachedExchangeRates?: ConversionRates) {
     console.log(`Creating new Stripe product for ${supabaseProduct.name}`);
     // Update price data for the product
     const unit_amount = Math.round(supabaseProduct.price*100)
     const lookup_key = `${supabaseProduct.sku}_${supabaseProduct.name}`
     const currency_options = await generateStripeCurrencyOptions(
         Math.round(supabaseProduct.price*100),
-        "PRODUCT"
+        "PRODUCT",
+        undefined,
+        cachedExchangeRates
     )
 
     process.stdout.write("Creating Stripe product... ")
@@ -207,7 +241,11 @@ async function createStripeProduct(supabaseProduct: ProductData) {
         description: supabaseProduct.description ?? undefined,
         name: supabaseProduct.name,
         images: imageURL ? [imageURL] : [],
-        metadata: {category_id: supabaseProduct.category.id, category: supabaseProduct.category.name},
+        metadata: {
+            category_id: supabaseProduct.category.id,
+            category: supabaseProduct.category.name,
+            sku: supabaseProduct.sku
+        },
         url: `https://thisshopissogay.com/products/${supabaseProduct.sku}`,
         default_price_data: { // Create new price
             currency: "gbp",
@@ -224,41 +262,6 @@ async function createStripeProduct(supabaseProduct: ProductData) {
         lookup_key, transfer_lookup_key: true
     })
     process.stdout.write("\r\x1b[K") // Clear line
-}
-
-/**
- * Update all products on Stripe based on all the products on Supabase
- * @param archiveAllOldPrices Whether to archive all the old prices, instead of just the current default, when updating
- * prices.
- */
-async function updateAllProducts(archiveAllOldPrices = false) {
-    // Fetch current products from Supabase & Stripe.
-    const supabaseProds = await getProducts(supabaseAnon);
-    const stripeProds = await fetchActiveStripeProducts();
-
-    // For each Stripe product, try to find a matching Supabase product to update information from.
-    for (const stripeProd of stripeProds) {
-        // If there's no associated metadata, the Stripe product is malformed and should be de-activated.
-        if (!stripeProd.metadata.sku) {await stripe.products.update(stripeProd.id, {active: false})}
-
-        // Search for matches
-        const matchedSupabaseProds = supabaseProds.filter(
-            supabaseProd => ""+supabaseProd.sku === stripeProd.metadata.sku)
-        // If a match is found, update Stripe and remove the product from the list of Supabase products
-        if (matchedSupabaseProds.length > 0) {
-            await updateStripeProduct(stripeProd, matchedSupabaseProds[0], archiveAllOldPrices)
-            supabaseProds.filter(p => p.sku !== matchedSupabaseProds[0].sku)
-        }
-        // If no match is found, disable this Stripe Product
-        else if (matchedSupabaseProds.length === 0) {
-            await stripe.products.update(stripeProd.id, {active: false})
-        }
-    }
-
-    // For each remaining Supabase product, create a new Stripe product since we know it doesn't exist yet.
-    for (const supabaseProd of supabaseProds) {
-        await createStripeProduct(supabaseProd);
-    }
 }
 
 /** Handle insertion of a new product record .
@@ -286,7 +289,23 @@ async function handleDeleteCase(prod: RawProductData) {
 async function handleUpdateCase(prod: RawProductData, archiveAllOldPrices = false) {
     const supabaseProds = await getProducts(supabaseAnon, [prod.sku]);
     const stripeProd = await fetchStripeProduct(prod.sku)
-    if (stripeProd) {
+    if (stripeProd && supabaseProds.length > 0) {
         await updateStripeProduct(stripeProd, supabaseProds[0], archiveAllOldPrices)
+    }
+}
+
+/**
+ * Update to product on Stripe from only a Supabase SKU. If no Stripe product exists, create one.
+ * @param sku
+ * @param archiveAllOldPrices
+ * @param cachedExchangeRates
+ */
+async function updateFromSku(sku: number, archiveAllOldPrices = false, cachedExchangeRates?: ConversionRates) {
+    const supabaseProds = await getProducts(supabaseAnon, [sku]);
+    const stripeProd = await fetchStripeProduct(sku);
+    if (stripeProd && supabaseProds.length > 0) {
+        await updateStripeProduct(stripeProd, supabaseProds[0], archiveAllOldPrices, cachedExchangeRates)
+    } else if (supabaseProds.length > 0) {
+        await createStripeProduct(supabaseProds[0], cachedExchangeRates)
     }
 }
